Apply the MVC pattern by moving calculation logic to the Model folder
Model receives data from the view(typically the user input) and do some logic. The outputs of the model are typically "get" functions. The get functions often reuturn data that the destination page wants.

By creating a seprate struct that holds all the data for Model outputs makes it easy to write the "get" functions.